Split job publishing into shared outbox transports and drop Job status.
