+++ apps/web/app/(ee)/api/cron/queue/retry/route.ts
@@ -1,6 +1,5 @@
 import { withCron } from "@/lib/cron/with-cron";
-import { publishPendingDefineJobs } from "@/lib/jobs";
-import { publishPendingWorkflows } from "@/lib/jobs/publish-workflows";
+import { publishPendingJobs } from "@/lib/jobs/outbox";
 import { redis } from "@/lib/upstash/redis";
 import { logAndRespond } from "../../utils";
 
@@ -12,8 +11,8 @@ export const dynamic = "force-dynamic";
 const LOCK_KEY = "lock:queue-retry";
 const LOCK_TTL_SECONDS = 600;
 
-// GET /api/cron/queue/retry – republish workflows / defineJob rows that failed
-// to publish to QStash at dispatch time; rows are deleted on successful publish
+// GET /api/cron/queue/retry – republish jobs that failed to publish to QStash
+// at dispatch time; rows are deleted on successful publish
 export const GET = withCron(async () => {
   const acquired = await redis.set(LOCK_KEY, "1", {
     nx: true,
@@ -27,14 +26,7 @@ export const GET = withCron(async () => {
   }
 
   try {
-    const [workflows, defineJobs] = await Promise.all([
-      publishPendingWorkflows(),
-      publishPendingDefineJobs(),
-    ]);
-
-    const attempted = workflows.attempted + defineJobs.attempted;
-    const published = workflows.published + defineJobs.published;
-    const failed = workflows.failed + defineJobs.failed;
+    const { attempted, published, failed } = await publishPendingJobs();
 
     if (attempted === 0) {
       return logAndRespond("No background jobs to retry.");
