+++ apps/web/lib/jobs/types.ts
@@ -1,0 +1,15 @@
+export type PublishResult = {
+  id: string;
+  status: "published" | "failed";
+  lastError: string | null;
+  messageId?: string;
+  workflowRunId?: string;
+};
+
+export type PersistableJob = {
+  id: string;
+  name: string;
+  payload: unknown;
+  options?: unknown;
+  scheduledAt?: Date;
+};
